% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.survivalfm}
\alias{predict.survivalfm}
\title{Predict using a survivalFM object}
\usage{
\method{predict}{survivalfm}(object, newx, type = "link")
}
\arguments{
\item{object}{Fitted "survivalfm" model object.}

\item{newx}{Matrix of new observations for which predictions are to be made.}

\item{type}{Type of prediction. Currently supports only "link", which gives the linear predictors.}
}
\value{
Returns a vector of the linear predictors.
}
\description{
Given a fitted survivalFM object, this function makes predictions on new observations.
}
