% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalfm.R
\name{survivalfm}
\alias{survivalfm}
\title{Fit survivalFM with specified parameters}
\usage{
survivalfm(
  x,
  y,
  rank,
  lambda1,
  lambda2,
  maxiter = 1000,
  reltol = sqrt(.Machine$double.eps),
  trace = 0,
  optimization_method = "BFGS"
)
}
\arguments{
\item{x}{Input data matrix; observations as rows, covariates/features as
columns.}

\item{y}{Response variable; \code{Surv} object from the survival package.}

\item{rank}{Rank of the factorization for the interaction terms.}

\item{lambda1}{Regularization strength for the linear terms.}

\item{lambda2}{Regularization strength for the factorized interaction terms.}

\item{maxiter}{Maximum number of iterations over the data for all
regularization parameter values. Default is 1000.}

\item{reltol}{Relative convergence tolerance for the optimization method in
\code{stats::optim}. Default is \code{sqrt(.Machine$double.eps)}.}

\item{trace}{If trace=1, will display messages of the progress.}

\item{optimization_method}{The optimization method used by
\code{stats::optim}. Default is "BFGS".}
}
\value{
Returns an object of class \code{"survivalfm"}. It is a list
containing the coefficients of the fitted model \code{beta} (linear effects)
and \code{P} (factorized interaction parameter matrix). Returns also the
approximated interaction effects in \code{PP}, given by the inner product
<p,p>. All model coefficients, including both the linear effects and
interaction effects are contained in \code{coefficients}.
}
\description{
This function fits a survivalFM model, given specified input parameters. If
you wish to automatically optimize the regularization parameters
\code{lambda1} and \code{lambda2}, use fit.survivalfm().
}
