% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.survivalfm.R
\name{val.survivalfm}
\alias{val.survivalfm}
\title{Fit survivalFM and optimize regularization parameters using a validation set}
\usage{
val.survivalfm(
  x,
  y,
  rank,
  frac = 0.2,
  nlambda = 20,
  lambda.min = 1e-04,
  lambda.max = 1,
  lambda_equal = FALSE,
  lambda1_range = NULL,
  lambda2_range = NULL,
  early_stopping = TRUE,
  val_idx = NULL,
  maxiter = 1000,
  reltol = sqrt(.Machine$double.eps),
  parallel = FALSE,
  trace = 0,
  optimization_method = "BFGS"
)
}
\arguments{
\item{x}{Input data matrix; observations as rows, covariates/features as
columns.}

\item{y}{Response variable; \code{Surv} object from the survival package.}

\item{rank}{Rank of the factorization for the interaction terms.}

\item{frac}{Fraction of data to use for validation data set. Default is 0.2
(20\%).}

\item{nlambda}{Number of values for \code{lambda1} and \code{lambda2} to
search in the logarithmic space between \code{lambda.min} and
\code{lambda.max}. If \code{nlambda} is set to its default value of 20, the
function will automatically generate 20 logarithmically spaced values for
each lambda within the specified range. Alternatively, one can provide
custom sequences for \code{lambda1} and \code{lambda2} using the
\code{lambda1_range} and \code{lambda2_range} parameters.}

\item{lambda.min}{The minimum boundary of the range within which the
regularization parameters \code{lambda1} and \code{lambda2} are searched.
This is the lower limit for generating lambda values unless a custom range
is provided. It must be a positive number.}

\item{lambda.max}{The maximum boundary of the range within which the
regularization parameters \code{lambda1} and \code{lambda2} are searched.
This is the upper limit for generating lambda values unless a custom range
is provided. It must be a positive number greater than \code{lambda.min}.}

\item{lambda_equal}{Defines whether only equally strong regularization for
the linear and the interaction part should be considered, i.e. set
\code{lambda1} = \code{lambda2}. Considering only equal values for the two
regularization parameters reduces the number of potential parameter
combinations to search and can therefore speed up the optimization process.
However, in some cases it might be desired to allow differing
regularization strengths for the linear and the interaction part, in which
case this should be set to \code{FALSE}. Default is \code{FALSE}.}

\item{lambda1_range}{Custom range of values for the regularization parameter
\code{lambda1} to be optimized (regularization for the linear terms;
optional). By default, the function will create 50 equally spaced values in
a logarithmic scale between 1 and 0.0001.}

\item{lambda2_range}{Custom range of values for the regularization parameter
\code{lambda2} to be optimized (regularization for the factorized
interaction terms; optional).}

\item{early_stopping}{Stop early if further decreasing lambda2 does not
improve performance (default = TRUE).}

\item{val_idx}{Row indices to be used as the validation set. If not set, the
function will randomly sample a validation set from the training data.}

\item{maxiter}{Maximum number of iterations over the data for all
regularization parameter values. Default is 1000.}

\item{reltol}{Relative convergence tolerance for the optimization method in
\code{stats::optim}. Default is \code{sqrt(.Machine$double.eps)}.}

\item{parallel}{If TRUE, use parallel foreach to fit the models with
different regularization strengths. The parallel backend must be registered
beforehand. Default is \code{FALSE}.}

\item{trace}{If trace=1, will display messages of the progress.}

\item{optimization_method}{The optimization method used by
\code{stats::optim}. Default is "BFGS".}
}
\value{
Returns an object of class \code{"val.survivalfm"}. It is a list
containing the coefficients of the final fitted model \code{beta} (linear
effects) and \code{P} (factorized interaction parameter matrix). Returns
also the interaction effects in \code{PP}, given by the inner product
<p,p>. All model coefficients, including both the linear effects and
interaction effects are contained in \code{coefficients}. The validation
results obtained for different values of the regularization parameters are
provided in \code{val.results}.
}
\description{
This function automatically optimizes the regularization parameters
\code{lambda1} and \code{lambda2} for survivalFM. It automatically partitions
the input dataset into a training set and a validation set.
}
\details{
The optimization process seeks to identify the values of \code{lambda1} and
\code{lambda2} that maximize the concordance index (C-index) on the
validation set. After determining the optimal regularization parameters, the
function retrains the model on the entire dataset using these parameters.

By default, the function generates validation sets randomly, ensuring that
events are evenly distributed between the validation and training sets.

Note that this function does not search for the optimal rank of the
factorization \code{k} and it needs to be provided by the user. It is
recommended to start with low values for \code{k} (e.g. 5 or 10).
}
